Log browser improvements.
The user can now set {x,y,z}-axis name, unit, and data scaler as function arguments. The benefit is that the Log Browser displays human-readable quantities. And, due to questionable Labber interface choices, a user could previously not change the name of the Y-axis during image export -- which is remedied in this commit.