Included user copy of the CC-BY-NC-SA-4.0 license.
The file labelled LICENSE lists my terms for using this software. The CC-BY-NC-SA-4.0 header in all of my files in this repository is nontheless just as valid (as has been repeatedly shown in CC court cases) since they have all always included the link to the CC license. And the courts rule repeatedly that it is reasonable to assume that you understood that this was the license to the software.