Updated log export return for automation.
The functions, as well as the log exporter, will now return a list of strings to the master script that is running the entire system. These strings are file paths to the stored .h5 data files. The master script may take the files pointed to in these strings, and run suitable fitting routines on the data contained within. In short, the functions have been made more automatable.