Updated imaginary+real plane readout.
All data is now acquired as complex by default. Likewise, I have added a function that acquires qubit states |g⟩, |e⟩ and |f⟩ when reading out in either of those states. To make this work with the Presto (and the Labber Log Browser), I had to add an explicit force flag to the Data Exporter - where I user can force-flip whether to store columns or rows as data entries in the Log Browser.