Added major framework for state discrimination.
The state discrimination is not finished, although the algorithm is laid out and most of the needed functions are in place. Once unique n-qubit states (with m states considered in the state discrimination) can be represented as integers, one merely has to count the occurrences of integers along the depth of the discretised_data volume. This counting will yield the probability matrix for some user-provided state to look for.