Bug fix in state probability calculation.
The "n-ary" (binary, trinary) to decimal conversion in the data exporter would always be incorrect if the user's data did not happen to discriminate to the highest state present in the system. Example: if the user was looking for the state population of |01⟩, and had characterised the complex coordinate centre for state |2⟩ in the system, but no readout got state discriminated into |2⟩, then all of the state probability calculations for states |n0⟩, |n1⟩, and |n2⟩ would be invalid for all number strings n.