Added a "time remaining" printer.
The user inserts some number of seconds into the printer, and the appropriate amount of weeks, days, hours, minutes and seconds, gets printed in human-readable form. This function is useful for estimating the "true" length of a measurement that relies on uploading/downloading data from the Presto instrument between data points (but, while in the same experiment run).