Bug fix related to selecting which weighed data to be exported.
When determining which readout metrics to add to the .hdf5 logs, the wrong matrix was checked when trying to determine which metrics to send off to the data exporter. Example: if the user set weights 0.0 to the resulting area spanned in the complex plane by the analysed states, then the area would *as intended* not be added to the processed_data matrix that gets sent to the data exporter. However, the (now fixed) bug meant that all other metrics other than area would now never be added to the processed_data matrix, and thus never be put in the save file, due to an if-case that never triggered.