Added time trace plotting for |g⟩, |e⟩ and |f⟩. And finished the readout fidelity calculator.
The user can now plot the time traces for |g⟩, |e⟩ and |f⟩. When optimising the integration window start- and stop times, the user can opt-in to get fidelity plots. Thus the user can choose an integration window that instead optimises for the highest readout fidelity, instead of merely the area (or mean distance, or perimeter) spanned by all readout states in the complex plane.