Bug fix for single-entry shot_arr arrays.
If the user ran a single shot single-shot optimisation measurement (so, only a single measurement point in the shot_arr array), the optimiser would crash. Because, the Data exporter exports single-entry values as h5py attributes. This latter behaviour has not been changed for the Data exporter, the optimiser insteads creates its own shot_arr if the array is illegal.