Changes to state discrimination; a function for arbitrary state preparation and state discrimination thereof; and other minor updates.
The arbitrary state preparation file currently only supports |11⟩ and |22⟩. There are some listed changes to be done in this file. For instance, any and all measurements should never change the type of single-sideband mixing depending on what state is being prepared. This fact in fact goes for all measurement files, meaning that there is a phase-changing change planned for all measurement files. This change is necessary due to inconsistent complex plane readout of various state spaces, if the change is not made.