Added the ability to distinguish between first vs. second swept key in the exported H5PY files.
When stitching several exported files, one would loose the ext_keys order. The solution was to denote which key was the first one, and which one was the second one. The stitcher, which receives all keys alphabetically regardless, can now detect whether a key must be moved to the front of the swept_content array. This update allows the operator to give arbitrary names to the Log Browser export axes, without crashing the stitch function.