Added CZ local phase calibration using virtual-Z gates.
Only one gate envelope is needed. A frequency LUT is filled some IF frequency f with 0 → 2π for group 0. For this trick to work, f must be 0 for all entries. However, the USB mixing phase_q parameter will still change the output of the device, this factoid has been very confirmed. The user updates some phase onto some port, and a select_frequency call will alongside with some phase calculation functions calculate an appropriate (legal) phase to change to. This method of working with virtual-Z gates means that the Presto device's virtual-Z gates get a phase resolution of 12.27 mrad (2π / 512 entries).