Add simple fitting routine for finding resonator frequencies.
This fitting routine can fit the resonator trace to a Lorentzian. However, such fits are usually very shoddy, since realistic QPU resonators don't look like Lorentzians. Instead, the default for the routine is to record where the Lorentzian dips, and then instead just filter the resonator trace. The dip of the filtered resonator trace is much closer to the value for the resonator that you'd grab in the actual laboratory. The difference between this filtered trace dip, and the Lorentzian dip, is reported as the error bar for the value. Which in turn, in my testing, has proven to be quite a decent error bar estimate. Assuming that one were to only look at the magnitude plot for establishing a final value for "the resonator frequency."