Add fit script for two-tone spectroscopy.
The two-tone spectroscopy fitter attempts to fit a rudimentary spectral transition to a Gaussian distribution. If the input data is very noisy, then the fitter would likely be able to fit the data. However, the expectation value of the Gaussian will simply latch onto the highest peak in the dataset. The error bars are luckily enormous beyond belief, thus it's likely possible to catch a poor fit. I have verified that if the two-tone spectrography looks like a Gaussian in the spectrum, then the fitter works well.