Add utilities for DC biasing.
One new function now automatically calculates the to-be-provided integer for the DC DACs' range_i settings. This integer was erroneously set previously for certain user bias sweeps. The other new function detects the Presto API version, and "safely" sets the DC bias voltage of several (possibly interconnected) DC ports, depending on the detected Presto version. If the user's Presto API version is not detected among the list of "old versions," it is assumed that the user is using a modern API version.