Change input sanitisation for DC biasing.
All scripts that utilise DC biasing for a SQUID coupler, has had its user input argument sanitisation changed. The lead principle is to not replicate the same snippet of code in every file, but trying to consolidate the DC bias input argument checking into a seperate function in another file. As a side-effect, there is also no more exception handling for figuring out whether to include (or exclude) an exported string that indicates the type of DC biasing used during the measurement. Finally, functions labelled as deprecated in the Rabi oscillation file, have been removed as well.