Apply major changes to integration window sweeps.
When sweeping the integration windows, the population centres would change for a given integration interval. The commit ensures that new population centres for state discrimination are found for every pixel datapoint in the integration sweep optimisation plot. The commit also makes a minor aesthetic change to the "Estimated time remaining" printout.