Remove all instances of making IF arguments positive-only.
The data exporter has been updated to enable negative IF frequencies when post-processing readout data. Not that this commit constitutes a breaking change: previous measurement files where the raw time trace data was stored, will be exported incorrectly following this commit. The solution is to add a minus sign before one variable at a designated location inside of the post-process function inside of the data exporter, in case you still have to re-export old measurement data.