Enable analysis of bias and power sweeps that also sweep the IF.
A side-effect of the current implementation is that all IF sweeps must happen as the outermost loop of the measurement. I do not like this fact, meaning that there will probably be some input in the future where it was denoted to the post-processing function whether the IF frequency sweep happened in the inner or outer loop.