Finish function for measuring a SQUID's AC flux bias.
The function can be used to at least get an estimate for converting between your native amplitude units, and the AC flux bias that is appended onto a SQUID. In the case of the Presto, your native amplitude unit for some two-qubit gate, would be provided in full-scale units (FS). Whereas physicists generally (insert a very-Christian term that I do not want indexed by the Google crawler) if they see such a unit. Thus it is better and more comparable if you convert the amplitude axis of your two-qubit gate tuning plots, into a unit range of AC-Φ. That way, one can also take a look at your plot and determine how hard you are driving the system, in units that are comparable between different quantum processor architectures.