Add bias sweep methods that force the DC bias sweep to be in the outer measurement loop.
Due to inadequacies in the IMP API, using the biasing card on the Presto AWG lock-in amplifier while requiring the DC bias sweep to be done in the outer loop of the measurement, requires placing the entire measurement inside of one iteration of the sequencer. Meaning that the repeat functionality cannot be utilised. This is quite silly.