Change default measurement inner vs. outer loop order.
Due to the way that the IMP API is done, sweeping a coupler's DC bias as the inner loop of a measurement, is very beneficial in terms of instrument resources. However, this approach of modifying the DC voltage applied onto a SQUID, is detrimental to the expected behaviour of the SQUID. The best DC bias voltage change, is a change that is never done.