Implement DC bias ramping.
When initialising the DC biasing of a measurement, the code framework will now in a rather clever way move the DC bias from where it is currently located, to the initial DC biasing spot of the measurement. The move is typically done slowly, for now at a rate of 20 mV/s. The goal is to be gentle with the DC biasing of the SQUID. The bias calculator takes into account all of the range changes that are necessary to accomplish the slow DC bias sweep. As of writing, de-initialising the DC bias in a safe fashion, is not supported.