Add function which plots the whole logic table for iSWAP gates, CZ gates and composed SWAP sequences.
It is assumed that the user is strictly running 2-qubit gates. The user may themselves specify the names of the parameters that in turn specify the axes that were using in Labber during the measurement. The logic table plotter assumes a data structure that is identical to the one made y the Labber Log Browser.