Add support for .h5 files in the T₁, T₂*, T₂e fitter.
Major support upgrade, the user may now perform the T₁, T₂*, T₂e sequence fit, on files taken with the Indecorum backend. That is, regular .h5 files. To do this, in the spirit of Indecorum itself, provide a list of filepaths to these .h5 files. The rest of the fitting routine is the same as before.